+++ flink-python/src/main/java/org/apache/flink/table/runtime/runners/python/PythonScalarFunctionRunner.java
@@ -16,11 +16,14 @@
  * limitations under the License.
  */
 
-package org.apache.flink.table.functions.python;
+package org.apache.flink.table.runtime.runners.python;
 
 import org.apache.flink.annotation.Internal;
 import org.apache.flink.python.PythonFunctionRunner;
 import org.apache.flink.table.functions.ScalarFunction;
+import org.apache.flink.table.functions.python.PythonEnv;
+import org.apache.flink.table.functions.python.PythonFunctionInfo;
+import org.apache.flink.table.runtime.typeutils.BeamTypeUtils;
 import org.apache.flink.table.types.logical.RowType;
 import org.apache.flink.types.Row;
 
@@ -41,8 +44,8 @@ public class PythonScalarFunctionRunner extends AbstractPythonScalarFunctionRunn
 		PythonEnv pythonEnv,
 		RowType inputType,
 		RowType outputType,
-		String tempDir) {
-		super(taskName, resultReceiver, scalarFunctions, pythonEnv, inputType, outputType, tempDir);
+		String[] tempDirs) {
+		super(taskName, resultReceiver, scalarFunctions, pythonEnv, inputType, outputType, tempDirs);
 	}
 
 	@Override
