+++ flink-python/src/test/java/org/apache/flink/table/functions/python/BaseRowPythonScalarFunctionRunnerTest.java
@@ -19,7 +19,9 @@
 package org.apache.flink.table.functions.python;
 
 import org.apache.flink.table.dataformat.BaseRow;
-import org.apache.flink.table.functions.python.coders.BaseRowCoder;
+import org.apache.flink.table.runtime.runners.python.AbstractPythonScalarFunctionRunner;
+import org.apache.flink.table.runtime.runners.python.BaseRowPythonScalarFunctionRunner;
+import org.apache.flink.table.runtime.typeutils.coders.BaseRowCoder;
 import org.apache.flink.table.types.logical.RowType;
 
 import org.apache.beam.sdk.coders.Coder;
